Erase instead of mkfs'ing during format
Pong has "formattable" flag on both userdata and metadata, which lets us
to allow the system to format instead of using whatever mkfs logic that
Google ships in fastboot, which can be quite out-of-sync of what the
manufacturers intended.

With this, fastboot now only zeroes the partitions and the system detects
it's been erased and performs format with its mkfs.

Signed-off-by: Juhyung Park <qkrwngud825@gmail.com>